requirements.txt
pandas==1.4.4
geopy==2.2.0
pyyaml==6.0